Yices.Easy.Build

Description

Provides a monad transformer for constructing Yices queries.

This is not an essential part of the interface. Expressions may be built using Yices.Easy.Types directly.

The monad is simply a convenient way to handle bookkeeping regarding fresh names, accumulated declarations / assertions, and choosing variables to retrieve from the Model.

The monad transformer

data BuildT m a

Monad transformer for building Yices queries.

type Build a = BuildT Identity a

runBuildT :: Monad m => BuildT m a -> m (a, Query)

execBuildT :: Monad m => BuildT m a -> m Query

runBuild :: Build a -> (a, Query)

execBuild :: Build a -> Query

Invoking Yices directly

solveBuildT :: BuildT IO a -> IO Result

Invoke solve on a BuildT IO action directly.

solveBuild :: Build a -> IO Result

Invoke solve on a Build action directly.

Fresh names

freshName :: Monad m => Ident -> BuildT m Ident

Generate a fresh name from a base string.

The name will not conflict with a user-specified name unless that name begins with an underscore.

Declarations

declare :: Monad m => Ident -> Type -> BuildT m Expr

Declare a variable.

fresh :: Monad m => Ident -> Type -> BuildT m Expr

Declare a fresh variable given a base name.

See freshName for naming details.

Assertions

assert :: Monad m => Expr -> BuildT m ()

Assert a Boolean truth.

Getting variables

get :: Monad m => Ident -> VarType -> BuildT m ()

Arrange to get this variable in the resulting Model.

Declaring and getting variables

declInt :: Monad m => Ident -> BuildT m Expr

Declare an int variable and get it.

declBool :: Monad m => Ident -> BuildT m Expr

Declare a bool variable and get it.

declBitvec :: Monad m => Size -> Ident -> BuildT m Expr

Declare a bitvector variable and get it.