Yices.Easy.Sugar

Description

Syntactic sugar for embedding Yices expressions in Haskell.

Defines a number of aliases and infix operators, and a Num instance for Expr.

This is not an essential part of the interface. Expressions may be built using Yices.Easy.Types directly.

This is a quick hack. There are several other projects working on vastly more sophisticated SMT EDSLs, so this one is unlikely to receive much attention. That said, suggestions and patches are always welcome.

Types

real :: Type

int :: Type

nat :: Type

bool :: Type

Booleans

true :: Expr

false :: Expr

(||.) :: Expr -> Expr -> Expr

(&&.) :: Expr -> Expr -> Expr

Application

($.) :: Expr -> [Expr] -> Expr

Scalar arithmetic

We also provide instance Num Expr, so you can use (+), (*), (-), negate, fromInteger, and integer literals.

add :: [Expr] -> Expr

mul :: [Expr] -> Expr

sub :: [Expr] -> Expr

Scalar comparison

(==.) :: Expr -> Expr -> Expr

(/=.) :: Expr -> Expr -> Expr

(>.) :: Expr -> Expr -> Expr

(>=.) :: Expr -> Expr -> Expr

(<.) :: Expr -> Expr -> Expr

(<=.) :: Expr -> Expr -> Expr

Literal bitvectors

bv :: Size -> Integer -> Expr

bv s n has value n and size at least s.

Bitvector arithmetic

(+@) :: Expr -> Expr -> Expr

(-@) :: Expr -> Expr -> Expr

(*@) :: Expr -> Expr -> Expr

Signed bitvector comarpsion

(>@) :: Expr -> Expr -> Expr

(>=@) :: Expr -> Expr -> Expr

(<@) :: Expr -> Expr -> Expr

(<=@) :: Expr -> Expr -> Expr

Unsigned bitvector comparison

(>@@) :: Expr -> Expr -> Expr

(>=@@) :: Expr -> Expr -> Expr

(<@@) :: Expr -> Expr -> Expr

(<=@@) :: Expr -> Expr -> Expr

Bitvector concatenation

(++@) :: Expr -> Expr -> Expr

Bitwise operations

(|@) :: Expr -> Expr -> Expr

(&@) :: Expr -> Expr -> Expr

(^@) :: Expr -> Expr -> Expr