Yices.Easy.Types

Description

Types for Yices syntax, in the subset we support.

Top level

data Declaration

Declare the existence of a variable.

Constructors

Declare Ident Type

newtype Assertion

Assert a Boolean fact.

Constructors

Assert Expr

data Context

Assert some facts in the context of some variables.

Constructors

Context [Declaration] [Assertion]

Expressions

data Type

Types.

Constructors

TyName Ident
Named type.

TyFun [Type] Type
Function, with args and result type.

TyBitvec Size
Bitvector.

TyTuple [Type]
Tuple.

data Expr

Expressions.

Constructors

LitBool Bool
Literal boolean.

LitNum LitNum
Literal number.

Var Ident
Variable.

Apply Expr [Expr]
Function application.

Arith Arith [Expr]
Arithmetic.

Logic Logic [Expr]
Boolean logic operation.

Not Expr
Boolean negation.

Compare Compare Expr Expr
Comparison.

IfThenElse Expr Expr Expr
Conditional expression.

LitBitvec LitBitvec
Literal bitvector.

BitArith Arith Expr Expr
Bitvector arithmetic.

BitMinus Expr
Bitvector arithmetic negation.

BitConcat Expr Expr
Concatenate bitvectors.

BitLogic Logic Expr Expr
Bitwise logical operation.

BitNot Expr
Bitwise negate.

BitShift Direction Bit Size Expr
Bit shift with fill.

BitCompare Signedness Compare Expr Expr
Bitvector comparison.

BitExtract Index Index Expr
Extract a subvector.

BitSignEx Size Expr
Sign-extend n extra bits.

data LitNum

Ways of specifying a literal number.

Constructors

FromInt Int

FromString String
ASCII digits.

data LitBitvec

Ways of specifying a literal bitvector.

Constructors

FromULong Size CULong

FromBits [Bit]
Least-significant first.

data Arith

Arithmetic operations.

Constructors

Add

Sub

Mul

data Logic

Logical operations.

Constructors

And

Or

Xor
Bitvector only; a runtime error for scalars!

data Compare

Comparison operations.

Constructors

Eq

Ne

Gt

Ge

Lt

Le

data Direction

Direction of shifting.

Constructors

L

R

data Bit

Bits.

Constructors

B0

B1

data Signedness

Signedness of bitvector operations.

Constructors

Signed

Unsigned

Type synonyms

type Ident = String

type Size = Int

type Index = Int

Models

data VarType

Types of variables which can be queried from a model.

Constructors

VarBool

VarInt

VarRational

VarDouble

VarBitvec Size

data Get

A variable to query from the model.

Constructors

Get Ident VarType

type ModelType = [Get]

data Query

A query to the solver: solve for some variables in a context.

Constructors

Query Context ModelType

data Value

A value produced from a model.

Constructors

ValBool Bool

ValInt CLong

ValRational Rational

ValDouble CDouble

ValBitvec [Bit]
Least-significant first.

type Model = Map Ident Value

A model maps variables to values.

Some of the asked-for variables may be absent, if there was an error or the model was underspecified.

data Result

Result of trying to solve a model.

Constructors

Sat Model
Definitely satisfiable.

Unsat
Definitely unsatisfiable.

Unknown
Unknown, due to incompleteness or error.